requirements.txt
pillow==12.1.1
pystray==0.19.5